reactor: RealmManager 객체생성, PopoverView 파일 분리
